CMIS-1048: Workbench: fixed delete on a folder
git-svn-id: https://svn.apache.org/repos/asf/chemistry/opencmis/trunk@1815011 13f79535-47bb-0310-9956-ffa450edef68

+++ chemistry-opencmis-workbench/chemistry-opencmis-workbench/src/main/java/org/apache/chemistry/opencmis/workbench/worker/DeleteWorker.java
@@ -34,6 +34,7 @@ public class DeleteWorker extends InfoWorkbenchWorker {
     private boolean allversions;
     private UnfileObject unfile;
     private boolean continueOnFailure;
+    private boolean deleteTree;
 
     // out
     private List<String> failedIds;
@@ -49,6 +50,7 @@ public class DeleteWorker extends InfoWorkbenchWorker {
         this.allversions = allversions;
         this.unfile = UnfileObject.DELETE;
         this.continueOnFailure = true;
+        this.deleteTree = false;
     }
 
     public DeleteWorker(Component comp, Folder folder, boolean allversions, UnfileObject unfile,
@@ -59,6 +61,7 @@ public class DeleteWorker extends InfoWorkbenchWorker {
         this.allversions = allversions;
         this.unfile = unfile;
         this.continueOnFailure = continueOnFailure;
+        this.deleteTree = true;
     }
 
     @Override
@@ -73,7 +76,7 @@ public class DeleteWorker extends InfoWorkbenchWorker {
 
     @Override
     protected Object doInBackground() throws Exception {
-        if (cmisObject instanceof Folder) {
+        if (cmisObject instanceof Folder && deleteTree) {
             failedIds = ((Folder) cmisObject).deleteTree(allversions, unfile, continueOnFailure);
         } else {
             cmisObject.delete(allversions);
